% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{lwgeom_extSoftVersion}
\alias{lwgeom_extSoftVersion}
\title{Provide the external dependencies versions of the libraries linked to sf}
\usage{
lwgeom_extSoftVersion()
}
\description{
Provide the external dependencies versions of the libraries linked to sf
}
