% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid.R
\name{lwgeom_make_valid}
\alias{lwgeom_make_valid}
\title{Make an invalid geometry valid}
\usage{
lwgeom_make_valid(x)
}
\arguments{
\item{x}{object of class \code{sfc}}
}
\description{
Make an invalid geometry valid
}
