% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprec.R
\name{perimeter-deprecated}
\alias{perimeter-deprecated}
\alias{st_perimeter}
\title{compute perimeter from polygons or other geometries}
\usage{
st_perimeter(x)
}
\arguments{
\item{x}{object of class \code{sf}, \code{sfc} or \code{sfg}}
}
\description{
Deprecated.
Use \code{sf::st_perimeter()} or \code{st_perimeter_lwgeom()} instead.
}
\keyword{internal}
