% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perimeter.R
\name{perimeter}
\alias{perimeter}
\alias{st_perimeter}
\alias{perimeter}
\alias{st_perimeter_2d}
\title{compute perimeter from polygons or other geometries}
\usage{
st_perimeter(x)

st_perimeter_2d(x)
}
\arguments{
\item{x}{object of class \code{sf}, \code{sfc} or \code{sfg}}
}
\description{
compute perimeter from polygons or other geometries
}
