% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twkb.R
\name{st_as_sfc.TWKB}
\alias{st_as_sfc.TWKB}
\title{create sfc object from tiny well-known binary (twkb)}
\usage{
\method{st_as_sfc}{TWKB}(x, ...)
}
\arguments{
\item{x}{list with raw vectors, of class \code{TWKB}}

\item{...}{ignored}
}
\description{
create sfc object from tiny well-known binary (twkb)
}
\examples{
l = structure(list(as.raw(c(0x02, 0x00, 0x02, 0x02, 0x02, 0x08, 0x08))), class = "TWKB")
library(sf) # load generic
st_as_sfc(l)
}
\seealso{
https://github.com/TWKB/Specification/blob/master/twkb.md
}
