% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bearing.R
\name{st_geod_azimuth}
\alias{st_geod_azimuth}
\title{compute azimuth between sequence of points}
\usage{
st_geod_azimuth(x)
}
\arguments{
\item{x}{object of class \code{sf}, \code{sfc} or \code{sfg}}
}
\description{
compute azimuth between sequence of points
}
\examples{
library(sf)
p = st_sfc(st_point(c(7,52)), st_point(c(8,53)), crs = 4326)
st_geod_azimuth(p)
}
