% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subdivide.R
\name{st_subdivide}
\alias{st_subdivide}
\title{Return a collection of geometries resulting by subdividing a geometry}
\usage{
st_subdivide(x, max_vertices)
}
\arguments{
\item{x}{object with geometries to be subdivided}

\item{max_vertices}{integer; maximum size of the subgeometries (at least 8)}
}
\value{
object of the same class as \code{x}
}
\description{
Return a collection of geometries resulting by subdividing a geometry
}
\examples{
library(sf)
demo(nc, ask = FALSE, echo = FALSE)
x = st_subdivide(nc, 10)
plot(x[1])
}
