% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_x.R
\name{st_wrap_x}
\alias{st_wrap_x}
\title{Splits input geometries by a vertical line and moves components falling on
one side of that line by a fixed amount}
\usage{
st_wrap_x(x, wrap, move)
}
\arguments{
\item{x}{object with geometries to be split}

\item{wrap}{x value of split line}

\item{move}{amount by which geometries falling to the left of the line
should be translated to the right}
}
\value{
object of the same class as \code{x}
}
\description{
Splits input geometries by a vertical line and moves components falling on
one side of that line by a fixed amount
}
\examples{
library(sf)
demo(nc, ask = FALSE, echo = FALSE)
x = st_wrap_x(nc, -78, 10)
plot(x[1])
}
